utilities: Share Armed visualization between modules
