+++ src/utilities.h
@@ -1,0 +1,4 @@
+#pragma once
+#include <MicroBit.h>
+
+void DisplayArmed(bool isArmed, Image &ledDisplay);
